`aframe deploy`: port to use new CLI manager
