+++ commands/index.js
@@ -1,4 +1,3 @@
 module.exports = {
-  'deploy': require('./deploy.js'),
   'submit': require('./submit.js')
 };
