+++ lib/utils.js
@@ -8,7 +8,8 @@ const templatesJson = require('../templates/index.json');
 const pkgDefault = {
   scripts: {
     build: 'aframe build',
-    serve: 'aframe serve'
+    serve: 'aframe serve',
+    deploy: 'aframe deploy'
   }
 };
