+++ templates/aframe-model-template/package.json
@@ -8,7 +8,8 @@
   "scripts": {
     "start": "npm run serve",
     "serve": "aframe serve",
-    "build": "aframe build"
+    "build": "aframe build",
+    "deploy": "aframe deploy"
   },
   "dependencies": {
     "aframe": "^0.6.0",
